==> 2_clean_names.py <==
# Conversion by Huihao JING on October 19, 2021
# This code cleans the OpenSecrets and boardex datasets
# 1. Clean up employer/firm names - convert everything to upper case, get rid of 
  # punctuation, changing words such as corporation to corp, etc
# 2. For first and last names, also standardize. Get dictionary of first names 
  # to match nicknames to full names
import numpy as np
import datatable as dt
import os
import datetime
import pandas as pd
import pyreadr as pr
import datatable as dt
from datar.all import f, c, tibble, sum, group_by, mutate
start = datetime.datetime.now()
ddir = os.path.abspath("\home\ysun9\OpenSecrets")
rdir = os.path.join(ddir,"raw_data")
gdir = os.path.join(ddir,"generated_data")
# Read in data
'''
os_indiv <- os_indiv[, row := row.names(.SD), by = .(FirmID, ContribID)][row == 1][, row := NULL]
'''
os = pd.DataFrame(pr.read_r(gdir, r"open_secrets/opensecrets.RDS").keys('FirmID'))
os_indiv = os['Contrib','Firm','FirmID','Cycle','ContribID'][os['Firm'].notna()]
os_indiv['row'] = os_indiv['row']

os_indiv = os_indiv[['LastName','FirstName']] = os_indiv['Contrib'].str.split(",",expand=True,)
os_indiv['FirstName'] = os_indiv['FirstName'].str.lstrip()
os_indiv['FirstName'] = os_indiv['FirstName'].str.rstrip()
# Keep boardex members who have ever worked at a portfolio/institutional firm
'''boardex <- 
  readRDS(file.path(gdir, "BoardEx/boardex_empl.RDS")) %>%
  group_by(DirectorID, CompanyID) %>%
  filter(sum(Portfolio) > 0 | sum(Institutional) > 0) %>%
  mutate(DirectorName = str_replace_all(DirectorName, regex("Professor |Doctor "), ""), 
         LastName = word(DirectorName, -1), 
         FirstName = str_trim(str_replace(DirectorName, LastName, ""))) %>%
  ungroup  '''
boardex = pd.DataFrame(pr.read_r(gdir, "BoardEx/boardex_empl.RDS"))
boardex = boardex['DirectorID','CompanyID'][boardex['Portfolio'].sum > 0 | boardex['Institutional'].sum > 0]
boardex = boardex >> mutate(DirectorName = boardex['DirectorName'].str.replace("Professor |Doctor ", ""), 
         LastName = boardex['DirectorName'][-1],   
         FirstName = boardex['DirectorName'].str.replace(['LastName'],'').strip())     

# ==============================================================================
# Create a function to clean firms to apply to both OpenSecrets and BoardEx data
# ==============================================================================
def clean_firms(Firm):
  out = pd.DataFrame(Firm)
  out.loc['Firm'] = out['Firm'].str.replace('[^ -~]', "")
  out.loc['Firm'] = out['Firm'].str.replace("\\&|\\+",'',regex=True, inplace = True)
  out.loc['Firm'] = out['Firm'].str.replace("\\s\\(.*\\)", " ",regex=True, inplace = True)
  out.loc['Firm'] = out['Firm'].str.replace(r'[^\w\s]+', '')
  out.loc['temp'] = "!"
  out.loc['Firm'].str.cat(out['temp'])
  out.loc['Firm'] = out['Firm'].str.upper()
  out.loc['Firm'] = out['Firm'].str.strip()
  out.loc['Firm'] = out['Firm'].str.replace(" CAP MGMT", " CAPITAL MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" CAP ", " CAPITAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" GBL ", " GLOBAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMMERCIAL ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMMERCIA ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMMERCI ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMMERC ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMMER ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMME ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" COML ", " COMM ")
  out.loc['Firm'] = out['Firm'].str.replace(" SAVINGS ", " SVGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" SAVING ", " SVGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" SAVIN ", " SVGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" SAVI ", " SVGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" SAV ", " SVGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" STR ", "STREET")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISORY ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISOR ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISERS ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISERS", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISERY ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISORS ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace("ADVISORS", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVISER ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADV ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" ADVR ", " ADVS ")
  out.loc['Firm'] = out['Firm'].str.replace(" TECHNOLOGY", " TECH")
  out.loc['Firm'] = out['Firm'].str.replace(" TECHNOLOG", " TECH")
  out.loc['Firm'] = out['Firm'].str.replace(" TECHNOLO", " TECH")
  out.loc['Firm'] = out['Firm'].str.replace(" TECHNOL", " TECH")
  out.loc['Firm'] = out['Firm'].str.replace(" TECHNO", " TECH")
  out.loc['Firm'] = out['Firm'].str.replace(" TECHN", " TECH")
  out.loc['Firm'] = out['Firm'].str.replace(" INTERNATIONAL", " INTL")
  out.loc['Firm'] = out['Firm'].str.replace(" INTERNATION", " INTL")
  out.loc['Firm'] = out['Firm'].str.replace(" NATL ", " NATIONAL ")
  out.loc['Firm'] = out['Firm'].str.replace("1ST ", " FIRST ")
  out.loc['Firm'] = out['Firm'].str.replace("2ND ", " SECOND ")
  out.loc['Firm'] = out['Firm'].str.replace("3RD ", " THIRD ")
  out.loc['Firm'] = out['Firm'].str.replace("4TH ", " FOURT ")
  out.loc['Firm'] = out['Firm'].str.replace("5TH ", " FIFTH ")
  out.loc['Firm'] = out['Firm'].str.replace("6TH ", " SIXTH ")
  out.loc['Firm'] = out['Firm'].str.replace("7TH ", " SEVENTH ")
  out.loc['Firm'] = out['Firm'].str.replace("8TH ", " EIGTH ")
  out.loc['Firm'] = out['Firm'].str.replace("9TH ", " NINTH ")
  out.loc['Firm'] = out['Firm'].str.replace("FRST ", " FIRST ")
  out.loc['Firm'] = out['Firm'].str.replace(" VANCE SA ", " VANCE SANDERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" VANCE SND ", " VANCE SANDERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" VANCE SANDE ", " VANCE SANDERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" VANCE SANDR ", " VANCE SANDERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" SERVICE ", " SERVICES ")
  out.loc['Firm'] = out['Firm'].str.replace(" SVCS ", " SERVICES ")
  out.loc['Firm'] = out['Firm'].str.replace(" SVC ", " SERVICES ")
  out.loc['Firm'] = out['Firm'].str.replace(" SERV ", " SERVICES ")
  out.loc['Firm'] = out['Firm'].str.replace(" GRP ", " GROUP ")
  out.loc['Firm'] = out['Firm'].str.replace(" DEV ", " DEVELOPMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" DEVEL ", " DEVELOPMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" LIMITED PTNR ", " LP ")
  out.loc['Firm'] = out['Firm'].str.replace(" LIMITED PARTNERS ", " LP ")
  out.loc['Firm'] = out['Firm'].str.replace(" LIMITED PARTNER ", " LP ")
  out.loc['Firm'] = out['Firm'].str.replace(" NATL ", " NATIONAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" LIMITED ", " LTD ")
  out.loc['Firm'] = out['Firm'].str.replace(" N A ", " NA ")
  out.loc['Firm'] = out['Firm'].str.replace(" GBL ", " GLOBAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW YORK ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" CTY ", " CITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVT ", " INV ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVESTMENT ", " INV ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVESTOR ", " INV ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVEST ", " INV ")
  out.loc['Firm'] = out['Firm'].str.replace(" TR ", " TRUST ")
  out.loc['Firm'] = out['Firm'].str.replace(" BK ", " BANK ")
  out.loc['Firm'] = out['Firm'].str.replace(" BNK ", " BANK ")
  out.loc['Firm'] = out['Firm'].str.replace(" TRST", " TRUST")
  out.loc['Firm'] = out['Firm'].str.replace(" TRUST AND BANK", " BANK AND TRUST ")
  out.loc['Firm'] = out['Firm'].str.replace(" FND ", " FUND ")
  out.loc['Firm'] = out['Firm'].str.replace(" FD ", " FUND ")
  out.loc['Firm'] = out['Firm'].str.replace(" FUNDS ", " FUND ")
  out.loc['Firm'] = out['Firm'].str.replace(" BANCSHARES ", " BANKSHARES ")
  out.loc['Firm'] = out['Firm'].str.replace(" BCSHS ", " BANKSHARES ")
  out.loc['Firm'] = out['Firm'].str.replace(" BKSHS ", " BANKSHARES ")
  out.loc['Firm'] = out['Firm'].str.replace(" BCSHS ", " BANKSHARES ")
  out.loc['Firm'] = out['Firm'].str.replace(" PUB ", " PUBLIC ")
  out.loc['Firm'] = out['Firm'].str.replace(" A R ASSET MANAGEMENT", " AR ASSET MANAGEMENT")
  out.loc['Firm'] = out['Firm'].str.replace(" A I M ", " AIM ")
  out.loc['Firm'] = out['Firm'].str.replace(" FRST", " FIRST")
  out.loc['Firm'] = out['Firm'].str.replace(" FRST", " FIRST")
  out.loc['Firm'] = out['Firm'].str.replace(" WELFAR ", " WELFARE")
  out.loc['Firm'] = out['Firm'].str.replace(" W MUT LF ", " WELFARE MUTUAL LIFE")
  out.loc['Firm'] = out['Firm'].str.replace(" WELFARE MUT ", " WELFARE MUTUAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" WEL MUT ", " WELFARE MUTUAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" WELFARE MT ", " WELFARE MUTUAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" U S ", " US")
  out.loc['Firm'] = out['Firm'].str.replace(" United States", " US")
  out.loc['Firm'] = out['Firm'].str.replace(" RES AND MGMT", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RES AND MANAGE", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RESRCH AND MGMT", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RESR AND MGMT", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RESRCH AND MGMT", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RES AND MANAGEMENT", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RES AND MANAGEMEN", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGEMENT AND RESEARCH ", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MGMT AND RESEARCH ", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MGMT AND RES", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MGMT AND RESRCH", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RESEARCH AND MAN ", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" RES AND MAN ", " RESEARCH AND MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGEMENT", " MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGEMEN", " MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGEME", " MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGEM", " MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGE ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAG ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MGT ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MGM ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MNGMENT", " MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MNGMEN", " MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGERS", " MANAGER")
  out.loc['Firm'] = out['Firm'].str.replace(" INV MAN ", " INVESTMENT MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" INVESTMENT MAN ", " INVESTMENT MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" INVMT ", " INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" FNL ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FINL ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace("FINL ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FINANCIA ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FINANCI ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FINANC ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FINAN ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FINA ", " FINANCIAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" RET SYS", " RETIREMENT SYSTEM")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIR ", " RETIREMENT")
  out.loc['Firm'] = out['Firm'].str.replace(" RET SYSTEM ", " RETIREMENT SYSTEM")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIREMNT ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIREMEN ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIREME ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIREM ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIRE ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIRE ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIR ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" RETI ", " RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLOYEES RE", " EMP RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLOYEE RE", " EMP RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMP RET ", " EMP RETIREMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSURANCE ", " ASSUR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSURANCE ", " ASSUR ")
  out.loc['Firm'] = out['Firm'].str.replace(" HOLDING ", " HOLDINGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" HLDGS ", " HOLDINGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" HLDG ", " HOLDINGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" HLD ", " HOLDINGS ")
  out.loc['Firm'] = out['Firm'].str.replace(" LOEB HD", " LOEB HOLDINGS")
  out.loc['Firm'] = out['Firm'].str.replace(" BROS ", " BROTHERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" BRTHRS ", " BROTHERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" LOAN MTG", " LOAN MORTGAGE")
  out.loc['Firm'] = out['Firm'].str.replace(" RET AND EQUITIES", " RETIREMENT & EQUITIES")
  out.loc['Firm'] = out['Firm'].str.replace(" RET EQUITIES", " RETIREMENT & EQUITIES")
  out.loc['Firm'] = out['Firm'].str.replace(" RET AND EQTY", " RETIREMENT & EQUITIES")
  out.loc['Firm'] = out['Firm'].str.replace(" RET AND EQUITY", " RETIREMENT & EQUITIES")
  out.loc['Firm'] = out['Firm'].str.replace(" RET AND EQUITY", " RETIREMENT & EQUITIES")
  out.loc['Firm'] = out['Firm'].str.replace(" RETIREMENT EQ", " RETIREMENT & EQUITIES")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPORTUNITIES ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPORTUNIT ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPORTUNI ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPORTUN ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPORTU ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPORT ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPOR ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPO ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" OPPTY ", " OPPORTUNITY ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOVER ", " SOVEREIGN ")
  out.loc['Firm'] = out['Firm'].str.replace(" PRIVAT ", " PRIVATE ")
  out.loc['Firm'] = out['Firm'].str.replace(" PRIVA ", " PRIVATE ")
  out.loc['Firm'] = out['Firm'].str.replace(" PRIV ", " PRIVATE ")
  out.loc['Firm'] = out['Firm'].str.replace(" PLN ", " PLAN ")
  out.loc['Firm'] = out['Firm'].str.replace(" PLA ", " PLAN ")
  out.loc['Firm'] = out['Firm'].str.replace(" CHEM ", " CHEMICAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" CNDA ", " CANADA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CANA ", " CANADA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CANAD ", " CANADA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MUT LF ", " MUTUAL LIFE ")
  out.loc['Firm'] = out['Firm'].str.replace(" MUT LIFE ", " MUTUAL LIFE ")
  out.loc['Firm'] = out['Firm'].str.replace(" MUTUAL LF ", " MUTUAL LIFE ")
  out.loc['Firm'] = out['Firm'].str.replace(" MUT CAP ", " MUTUAL CAPITAL "),
  out.loc['Firm'] = out['Firm'].str.replace(" MUTL ", " MUTUAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" MUT ", " MUTUAL ")
  out.loc['Firm'] = out['Firm'].str.replace(" MUTUAL LF ", " MUTUAL LIFE ")
  out.loc['Firm'] = out['Firm'].str.replace(" LF ", " LIFE ")
  out.loc['Firm'] = out['Firm'].str.replace(" INS ", " INSURANCE ")
  out.loc['Firm'] = out['Firm'].str.replace(" INSUR ", " INSURANCE ")
  out.loc['Firm'] = out['Firm'].str.replace(" ANNTY ", " ANN ")
  out.loc['Firm'] = out['Firm'].str.replace(" ANNITY ", " ANN ")
  out.loc['Firm'] = out['Firm'].str.replace(" ANNIT ", " ANN ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSN ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCN ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCIATION ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASC", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASST MANA ", " ASSET MGMT")
  out.loc['Firm'] = out['Firm'].str.replace(" ASST ", " ASSET ")
  out.loc['Firm'] = out['Firm'].str.replace(" AMERICAN ", " AMER ")
  out.loc['Firm'] = out['Firm'].str.replace(" AMERICA ", " AMER ")
  out.loc['Firm'] = out['Firm'].str.replace(" AMERIC ", " AMER ")
  out.loc['Firm'] = out['Firm'].str.replace(" AMERIC ", " AMER ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLOYEES ", " EMP ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLOYEE ", " EMP ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLOYE ", " EMP ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLOY ", " EMP ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPLO ", " EMP ")
  out.loc['Firm'] = out['Firm'].str.replace(" EMPL ", " EMP ")
  out.loc['Firm'] = out['Firm'].str.replace(" SCH ", " SCHOOL ")
  out.loc['Firm'] = out['Firm'].str.replace(" DF ", " FUND ")
  out.loc['Firm'] = out['Firm'].str.replace(" SYS ", " SYSTEM ")
  out.loc['Firm'] = out['Firm'].str.replace(" BD ", " BOARD ")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGEMTENT", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MGNT ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MGN ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MGT ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MANAGE ", " MGMT ")
  out.loc['Firm'] = out['Firm'].str.replace(" PEN ", " PENSION ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENS ", " PENSION ")
  out.loc['Firm'] = out['Firm'].str.replace(" PNS ", " PENSION ")
  out.loc['Firm'] = out['Firm'].str.replace(" STS ", " STATES ")
  out.loc['Firm'] = out['Firm'].str.replace(" STL ", " STEEL ")
  out.loc['Firm'] = out['Firm'].str.replace(" STS ", " STATES ")
  out.loc['Firm'] = out['Firm'].str.replace(" SEC ", " SECURITIES ")
  out.loc['Firm'] = out['Firm'].str.replace(" SECS ", " SECURITIES ")
  out.loc['Firm'] = out['Firm'].str.replace(" FORGN ", " FOREIGN ")
  out.loc['Firm'] = out['Firm'].str.replace(" FORGN ", " FOREIGN ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCIATES ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCIATES", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCIATE ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCIAT ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCIA ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" ASSOCI ", " ASSOC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORPORATED ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORPORATE ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORPORAT ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORPORA ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORPOR ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORPO ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCORP ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCOR ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" INCO ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPORATION ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPORATIO ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPORATI ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPORAT ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPORA ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPOR ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CORPO ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" CRP ", " CORP ")
  out.loc['Firm'] = out['Firm'].str.replace(" L P", " LP")
  out.loc['Firm'] = out['Firm'].str.replace(" L L C", " LLC")
  out.loc['Firm'] = out['Firm'].str.replace(" L L P", " LLP")
  out.loc['Firm'] = out['Firm'].str.replace(" COMPANY", " CO")
  out.loc['Firm'] = out['Firm'].str.replace(" COMPAN ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMPA ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" COMP ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" COM ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" PTNRS ", " PARTNER ")
  out.loc['Firm'] = out['Firm'].str.replace(" PTNR ", " PARTNER ")
  out.loc['Firm'] = out['Firm'].str.replace(" PARTNERS ", " PARTNER ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVESTMENTS ", " INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVT ", " INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVMNT ", " INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" INVESTMENTS ", " INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" HOSPITAL ", " HOSP ")
  out.loc['Firm'] = out['Firm'].str.replace(" HOSPITA ", " HOSP ")
  out.loc['Firm'] = out['Firm'].str.replace(" HOSPIT ", " HOSP ")
  out.loc['Firm'] = out['Firm'].str.replace(" HOSPI ", " HOSP ")
  out.loc['Firm'] = out['Firm'].str.replace(" UNIVER ", " UNI ")
  out.loc['Firm'] = out['Firm'].str.replace(" UNIV ", " UNI ")
  out.loc['Firm'] = out['Firm'].str.replace(" UNIVERSITY ", " UNI ")
  out.loc['Firm'] = out['Firm'].str.replace("UNIVERSITY ", "UNI ")
  out.loc['Firm'] = out['Firm'].str.replace(" United States", " US")
  out.loc['Firm'] = out['Firm'].str.replace(" TEACHER ", " TEACHERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" TEACH ", " TEACHERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" GE INVESTMENT ", " GENERAL ELECTRIC INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" GE ASSET MANAGEMENT ", " GENERAL ELECTRIC ASSET MANAGEMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" GE INVESTMENT ", " GENERAL ELECTRIC INVESTMENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" GENERAL ELECTRI ", " GENERAL ELECTRIC")
  out.loc['Firm'] = out['Firm'].str.replace(" GENERAL ELECTR ", " GENERAL ELECTRIC")
  out.loc['Firm'] = out['Firm'].str.replace(" GENERAL ELECT ", " GENERAL ELECTRIC")
  out.loc['Firm'] = out['Firm'].str.replace(" GENERAL ELEC ", " GENERAL ELECTRIC")
  out.loc['Firm'] = out['Firm'].str.replace(" G E ", " GENERAL ELECTRIC")
  out.loc['Firm'] = out['Firm'].str.replace(" GENERAL RE ", " GENERAL REINSURANCE")
  out.loc['Firm'] = out['Firm'].str.replace(" STRATEGIC ", " STRAT")
  out.loc['Firm'] = out['Firm'].str.replace(" STRATEGI ", " STRAT")
  out.loc['Firm'] = out['Firm'].str.replace(" STRATEG ", " STRAT")
  out.loc['Firm'] = out['Firm'].str.replace(" STRATE ", " STRAT")
  out.loc['Firm'] = out['Firm'].str.replace(" STRATE ", " STRAT")
  out.loc['Firm'] = out['Firm'].str.replace(" FDTN ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" FOUNDATIO ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" FOUNDATI ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" FOUNDAT ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" FOUNDA ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" FOUND ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" FOUN ", " FOUNDATION ")
  out.loc['Firm'] = out['Firm'].str.replace(" MFRS ", " MANUFACTURERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MANUFACTURER ", " MANUFACTURERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MFG ", " MANUFACTURERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MFC ", " MANUFACTURERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MFG ", " MANUFACTURERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" HANOVER ", " HAN ")
  out.loc['Firm'] = out['Firm'].str.replace(" MERCHANT ", " MERCHANTS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MER ", " MERCHANTS ")
  out.loc['Firm'] = out['Firm'].str.replace(" BRITISH PET ", " BRITISH PETROLEUM ")
  out.loc['Firm'] = out['Firm'].str.replace(" BRITISH PETE ", " BRITISH PETROLEUM ")
  out.loc['Firm'] = out['Firm'].str.replace(" BRITISH PETRO ", " BRITISH PETROLEUM ")
  out.loc['Firm'] = out['Firm'].str.replace(" BRTS ", " BRITISH ")
  out.loc['Firm'] = out['Firm'].str.replace(" PETRO ", " PETROLEUM ")
  out.loc['Firm'] = out['Firm'].str.replace(" N Y ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" FROST BKS ", " FROST BANKERS ")
  out.loc['Firm'] = out['Firm'].str.replace(" VENTURES ", " VENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" VENTURE ", " VENT ")
  out.loc['Firm'] = out['Firm'].str.replace(" ENTERPRISES ", " ENTR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ENTERPRISE ", " ENTR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ENTERPRS ", " ENTR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALABAMA ", " AL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALABAM ", " AL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALABA ", " AL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALAB ", " AL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALASKA ", " AK ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALASK ", " AK ")
  out.loc['Firm'] = out['Firm'].str.replace(" ALAS ", " AK ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARIZONA ", " AZ ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARIZON ", " AZ ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARIZO ", " AZ ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARKANSAS ", " AR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARKANSA ", " AR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARKANS ", " AR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ARKAN ", " AR ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALIFORNIA ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALIFORNI ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALIFORN ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALIFOR ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALIFO ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALIF ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" CALI ", " CA ")
  out.loc['Firm'] = out['Firm'].str.replace(" COLORADO ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" COLORAD ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" COLORA ", " CO ")
  out.loc['Firm'] = out['Firm'].str.replace(" CONNECTICUT ", " CT ")
  out.loc['Firm'] = out['Firm'].str.replace(" CONNECTICU ", " CT ")
  out.loc['Firm'] = out['Firm'].str.replace(" CONNECTIC ", " CT ")
  out.loc['Firm'] = out['Firm'].str.replace(" CONNECTI ", " CT ")
  out.loc['Firm'] = out['Firm'].str.replace(" CONN ", " CT ")
  out.loc['Firm'] = out['Firm'].str.replace(" DELAWARE ", " DE ")
  out.loc['Firm'] = out['Firm'].str.replace(" DELAWAR ", " DE ")
  out.loc['Firm'] = out['Firm'].str.replace(" DELAWA ", " DE ")
  out.loc['Firm'] = out['Firm'].str.replace(" DELAW ", " DE ")
  out.loc['Firm'] = out['Firm'].str.replace(" FLORIDA ", " FL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FLORID ", " FL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FLORI ", " FL ")
  out.loc['Firm'] = out['Firm'].str.replace(" FLOR ", " FL ")
  out.loc['Firm'] = out['Firm'].str.replace(" GEORGIA ", " GA ")
  out.loc['Firm'] = out['Firm'].str.replace(" GEORGI ", " GA ")
  out.loc['Firm'] = out['Firm'].str.replace(" HAWAII ", " HI ")
  out.loc['Firm'] = out['Firm'].str.replace(" HAWAI ", " HI ")
  out.loc['Firm'] = out['Firm'].str.replace(" IDAHO ", " ID ")
  out.loc['Firm'] = out['Firm'].str.replace(" IDAH ", " ID ")
  out.loc['Firm'] = out['Firm'].str.replace(" ILLINOIS ", " IL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ILLINOI ", " IL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ILLINO ", " IL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ILLIN ", " IL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ILLI ", " IL ")
  out.loc['Firm'] = out['Firm'].str.replace(" ILL ", " IL ")
  out.loc['Firm'] = out['Firm'].str.replace(" INDIANA ", " IN ")
  out.loc['Firm'] = out['Firm'].str.replace(" IOWA ", " IA ")
  out.loc['Firm'] = out['Firm'].str.replace(" IOW ", " IA ")
  out.loc['Firm'] = out['Firm'].str.replace(" KANSAS ", " KS ")
  out.loc['Firm'] = out['Firm'].str.replace(" KANSA ", " KS ")
  out.loc['Firm'] = out['Firm'].str.replace(" KENTUCKY ", " KY ")
  out.loc['Firm'] = out['Firm'].str.replace(" KENTUCK ", " KY ")
  out.loc['Firm'] = out['Firm'].str.replace(" KENTCKY ", " KY ")
  out.loc['Firm'] = out['Firm'].str.replace(" LOUISIANA ", " LA ")
  out.loc['Firm'] = out['Firm'].str.replace(" LOUISIAN ", " LA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MAINE ", " ME ")
  out.loc['Firm'] = out['Firm'].str.replace(" MARYLAND ", " MD ")
  out.loc['Firm'] = out['Firm'].str.replace(" MARYLND ", " MD ")
  out.loc['Firm'] = out['Firm'].str.replace(" MARYLAN ", " MD ")
  out.loc['Firm'] = out['Firm'].str.replace(" MARYLA ", " MD ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACHUSETTS ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACHUSETT ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACHUSET ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACHUSE ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACHUS ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACHU ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSACH ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSAC ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASSA ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MASS ", " MA ")
  out.loc['Firm'] = out['Firm'].str.replace(" MICHIGAN ", " MI ")
  out.loc['Firm'] = out['Firm'].str.replace(" MICHIGA ", " MI ")
  out.loc['Firm'] = out['Firm'].str.replace(" MICHIG ", " MI ")
  out.loc['Firm'] = out['Firm'].str.replace(" MICHI ", " MI ")
  out.loc['Firm'] = out['Firm'].str.replace(" MINNESOTA ", " MN ")
  out.loc['Firm'] = out['Firm'].str.replace(" MINNESOT ", " MN ")
  out.loc['Firm'] = out['Firm'].str.replace(" MINNESO ", " MN ")
  out.loc['Firm'] = out['Firm'].str.replace(" MINNES ", " MN ")
  out.loc['Firm'] = out['Firm'].str.replace(" MINN ", " MN ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSISSIPPI ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSISSIPP ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSISSIP ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSISSI ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSISS ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSIS ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSI ", " MS ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSOURI ", " MO ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSOUR ", " MO ")
  out.loc['Firm'] = out['Firm'].str.replace(" MISSOU ", " MO ")
  out.loc['Firm'] = out['Firm'].str.replace(" MONTANA ", " MT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MONTAN ", " MT ")
  out.loc['Firm'] = out['Firm'].str.replace(" MONTA ", " MT ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEBRASKA ", " NE ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEBRASK ", " NE ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEBRAS ", " NE ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEBRA ", " NE ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEBR ", " NE ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEVADA ", " NV ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEVAD ", " NV ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW HAMPSHIRE ", " NH ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW HAMPSHIR ", " NH ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW HAMPSHI ", " NH ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW HAMPSH ", " NH ")
  out.loc['Firm'] = out['Firm'].str.replace(" NW HAMPSHIRE ", " NH ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW JERSEY ", " NJ ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW JERSE ", " NJ ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW JERS ", " NJ ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW JER ", " NJ ")
  out.loc['Firm'] = out['Firm'].str.replace(" NW JERSEY ", " NJ ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW MEXICO ", " NM ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW MEXIC ", " NM ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW MEXI ", " NM ")
  out.loc['Firm'] = out['Firm'].str.replace(" NW MEXICO ", " NM ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW YORK ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW YOR ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEW YO ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" NW YORK ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" NEWYORK ", " NY ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH CAROLINA ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH CAROLIN ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH CAROLI ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH CAROL ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH CARO ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NRTH CAROLI ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH CAROLI ", " NC ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH DAKOTA ", " ND ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH DAKOT ", " ND ")
  out.loc['Firm'] = out['Firm'].str.replace(" NORTH DAKO ", " ND ")
  out.loc['Firm'] = out['Firm'].str.replace(" NRTH DAKOTA ", " ND ")
  out.loc['Firm'] = out['Firm'].str.replace(" OHIO ", " OH ")
  out.loc['Firm'] = out['Firm'].str.replace(" OKLAHOMA ", " OK ")
  out.loc['Firm'] = out['Firm'].str.replace(" OKLAHOM ", " OK ")
  out.loc['Firm'] = out['Firm'].str.replace(" OKLAHO ", " OK ")
  out.loc['Firm'] = out['Firm'].str.replace(" OKLAH ", " OK ")
  out.loc['Firm'] = out['Firm'].str.replace(" OKLHMA ", " OK ")
  out.loc['Firm'] = out['Firm'].str.replace(" OREGON ", " OR ")
  out.loc['Firm'] = out['Firm'].str.replace(" OREGN ", " OR ")
  out.loc['Firm'] = out['Firm'].str.replace(" ORGN ", " OR ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSYLVANIA ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSYLVANI ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSYLVAN ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSYLVA ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSYLV ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSYL ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNSY ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNS ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENNA ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" PENN ", " PA ")
  out.loc['Firm'] = out['Firm'].str.replace(" RHODE ISLAND ", " RI ")
  out.loc['Firm'] = out['Firm'].str.replace(" RHODE ISLAN ", " RI ")
  out.loc['Firm'] = out['Firm'].str.replace(" RHODE ISLND ", " RI ")
  out.loc['Firm'] = out['Firm'].str.replace(" RHODE ISL ", " RI ")
  out.loc['Firm'] = out['Firm'].str.replace(" RHDE ISLAND ", " RI ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH CAROLINA ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH CAROLIN ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH CAROLI ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH CAROL ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH CARO ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOTH CAROLI ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH CAROLI ", " SC ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH DAKOTA ", " SD ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH DAKOT ", " SD ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH DAKO ", " SD ")
  out.loc['Firm'] = out['Firm'].str.replace(" SOUTH DAK ", " SD ")
  out.loc['Firm'] = out['Firm'].str.replace(" TENNESSEE ", " TN ")
  out.loc['Firm'] = out['Firm'].str.replace(" TENNESSE ", " TN ")
  out.loc['Firm'] = out['Firm'].str.replace(" TENNESS ", " TN ")
  out.loc['Firm'] = out['Firm'].str.replace(" TNNSSE ", " TN ")
  out.loc['Firm'] = out['Firm'].str.replace(" TNNSS ", " TN ")
  out.loc['Firm'] = out['Firm'].str.replace(" TEXAS ", " TX ")
  out.loc['Firm'] = out['Firm'].str.replace(" TXS ", " TX ")
  out.loc['Firm'] = out['Firm'].str.replace(" UTAH ", " UT ")
  out.loc['Firm'] = out['Firm'].str.replace(" UTH ", " UT ")
  out.loc['Firm'] = out['Firm'].str.replace(" VERMONT ", " VT ")
  out.loc['Firm'] = out['Firm'].str.replace(" VERMON ", " VT ")
  out.loc['Firm'] = out['Firm'].str.replace(" VERMNT ", " VT ")
  out.loc['Firm'] = out['Firm'].str.replace(" VRMNT ", " VT ")
  out.loc['Firm'] = out['Firm'].str.replace(" VIRGINIA ", " VA ")
  out.loc['Firm'] = out['Firm'].str.replace(" VIRGINI ", " VA ")
  out.loc['Firm'] = out['Firm'].str.replace(" VIRGINA ", " VA ")
  out.loc['Firm'] = out['Firm'].str.replace(" VRGINIA ", " VA ")
  out.loc['Firm'] = out['Firm'].str.replace(" VRGNIA ", " VA ")
  out.loc['Firm'] = out['Firm'].str.replace(" WASHINGTON ", " WA ")
  out.loc['Firm'] = out['Firm'].str.replace(" WASHINGTO ", " WA ")
  out.loc['Firm'] = out['Firm'].str.replace(" WASHINGT ", " WA ")
  out.loc['Firm'] = out['Firm'].str.replace(" WASHING ", " WA ")
  out.loc['Firm'] = out['Firm'].str.replace(" WSHNGTN ", " WA ")
  out.loc['Firm'] = out['Firm'].str.replace(" WEST VIRGINIA ", " WV ")
  out.loc['Firm'] = out['Firm'].str.replace(" WST VIRGINIA ", " WV ")
  out.loc['Firm'] = out['Firm'].str.replace(" WEST VIRGINI ", " WV ")
  out.loc['Firm'] = out['Firm'].str.replace(" WEST VRGINIA ", " WV ")
  out.loc['Firm'] = out['Firm'].str.replace(" WISCONSIN ", " WI ")
  out.loc['Firm'] = out['Firm'].str.replace(" WISCONSI ", " WI ")
  out.loc['Firm'] = out['Firm'].str.replace(" WISCONSN ", " WI ")
  out.loc['Firm'] = out['Firm'].str.replace(" WISCON ", " WI ")
  out.loc['Firm'] = out['Firm'].str.replace(" WYOMING ", " WY ")
  out.loc['Firm'] = out['Firm'].str.replace(" WYOMIN ", " WY ")
  out.loc['Firm'] = out['Firm'].str.replace(" WYOMI ", " WY ")
  out.loc['Firm'] = out['Firm'].str.replace(" WYMNG ", " WY ")
  out.loc['Firm'] = out['Firm'].str.replace(" 1ST GLOBAL", "FIRST GLOBAL")
  out.loc['Firm'] = out['Firm'].str.replace(" 1ST SOURCE", "FIRST SOURCE")
  out.loc['Firm'] = out['Firm'].str.replace("A R ASSET ", "AR ASSET ")
  out.loc['Firm'] = out['Firm'].str.replace(" LTD PRT ", " LP ")
  out.loc['Firm'] = out['Firm'].str.replace(" IN ", " INC ")
  out.loc['Firm'] = out['Firm'].str.replace("A I M ", "AIM ")
  out.loc['Firm'] = out['Firm'].str.replace(" STRATEGIES ", " STRAT ")
  out.loc['Firm'] = out['Firm'].str.replace(" AND JEN ", " AND JENRETTE ")
  out.loc['Firm'] = out['Firm'].str.replace(" AND JENN ", " AND JENRETTE ")
  out.loc['Firm'] = out['Firm'].str.replace(" AND JENRETT ", " AND JENRETTE ")
  out.loc['Firm'] = out['Firm'].str.replace(" LUFKIN JENRETT ", " LUFKIN AND JENRETTE ")
  out.loc['Firm'] = out['Firm'].str.replace(" LUFKIN JENRTTE ", " LUFKIN AND JENRETTE ")
  out.loc['Firm'] = out['Firm'].str.replace(" P A", " PA")
  out.loc['Firm'] = out['Firm'].str.replace(" K C", " KC")
  out.loc['Firm'] = out['Firm'].str.replace(" S A", " SA")
  out.loc['Firm'] = out['Firm'].str.replace(' LLC ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' LTD ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' CO ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' LP ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' INC ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' PTN ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' PLC ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' CORP ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' LT ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' LL ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' L ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' TN ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' T ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' C ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' P ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' S ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' W ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' I ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' II ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' LLP ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' MGMT ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(' INTL ', ' ')
  out.loc['Firm'] = out['Firm'].str.replace(" !", "")
  out.loc['Firm'] = out['Firm'].str.strip()
  return(out.loc['Firm'])
# ==============================================================================
# Create a function to clean first and last names
# ==============================================================================
def clean_names(names):
  Name = pd.DataFrame(names)
  Name.loc['Name'] = Name['Name'].str.strip()
  Name.loc['Name'] = Name['Name'].str.upper()
  Name.loc['Name'] = Name['Name'].str.replace(" MRS$| MR$| MS$| MD$","",regex=True, inplace = True)
  out = Name['Name',1].tolist()
  return(out)
# =========================================
# apply cleaning firms and names functions to the two datasets
# ========================================
os_indiv.loc['Firm_std'] = clean_firms(os_indiv['Firm'])
os_indiv.loc['First_std'] = clean_firms(os_indiv['FirstName'])
os_indiv.loc['Last_std'] = clean_firms(os_indiv['LastName'])
os_std = pd.DataFrame(os_indiv)
os_std.to_csv(gdir+r"open_secrets/opensecrets_names_std.RDS")
boardex.loc['CompanyName_std'] = clean_firms(boardex['CompanyName'])
boardex.loc['First_std'] = clean_firms(boardex['FirstName'])
boardex.loc['Last_std'] = clean_firms(boardex['LastName'])
boardex_std = pd.DataFrame(boardex)
boardex_std.to_csv(gdir+r"intermediate/boardex_std.RDS")
print("made it to end of cleaning")
print("successfully saved")
end = datetime.datetime.now()
print("Boardex Import Done In: ", pd.to_datetime(end) - pd.to_datetime(start))
